Split unittest.
