+++ python/paddle/fluid/tests/unittests/CMakeLists.txt
@@ -709,11 +709,12 @@ set_tests_properties(test_nearest_interp_v2_op PROPERTIES TIMEOUT 120)
 set_tests_properties(test_trilinear_interp_op PROPERTIES TIMEOUT 120)
 set_tests_properties(test_bicubic_interp_v2_op PROPERTIES TIMEOUT 120)
 set_tests_properties(test_gather_op PROPERTIES TIMEOUT 120)
+set_tests_properties(test_static_save_load PROPERTIES TIMEOUT 120)
 if (WIN32)
-    set_tests_properties(test_static_save_load PROPERTIES TIMEOUT 900)
+    set_tests_properties(test_static_save_load_large PROPERTIES TIMEOUT 900)
     set_tests_properties(test_paddle_save_load PROPERTIES TIMEOUT 250)
 else()
-    set_tests_properties(test_static_save_load PROPERTIES TIMEOUT 600)
+    set_tests_properties(test_static_save_load_large PROPERTIES TIMEOUT 600)
     set_tests_properties(test_paddle_save_load PROPERTIES TIMEOUT 150)
 endif()
 set_tests_properties(test_imperative_selected_rows_to_lod_tensor PROPERTIES TIMEOUT 120)
